[matchingImageCollection] another split functions in cpp
